Move the files from bigquery/ to google/cloud/bigquery/
