+++ google/cloud/bigquery/client.h
@@ -17,9 +17,9 @@
 
 #include <memory>
 
-#include "bigquery/connection.h"
-#include "bigquery/connection_options.h"
-#include "bigquery/version.h"
+#include "google/cloud/bigquery/connection.h"
+#include "google/cloud/bigquery/connection_options.h"
+#include "google/cloud/bigquery/version.h"
 #include "google/cloud/status_or.h"
 
 namespace bigquery {
