+++ google/cloud/bigquery/internal/bigquerystorage_stub.cc
@@ -18,11 +18,11 @@
 
 #include <grpcpp/create_channel.h>
 
-#include "bigquery/connection.h"
-#include "bigquery/connection_options.h"
-#include "bigquery/internal/bigquerystorage_stub.h"
-#include "bigquery/internal/stream_reader.h"
-#include "bigquery/version.h"
+#include "google/cloud/bigquery/connection.h"
+#include "google/cloud/bigquery/connection_options.h"
+#include "google/cloud/bigquery/internal/bigquerystorage_stub.h"
+#include "google/cloud/bigquery/internal/stream_reader.h"
+#include "google/cloud/bigquery/version.h"
 #include "google/cloud/grpc_utils/grpc_error_delegate.h"
 #include "google/cloud/optional.h"
 #include "google/cloud/status_or.h"
