+++ google/cloud/bigquery/internal/connection_impl.cc
@@ -17,9 +17,9 @@
 #include <sstream>
 #include <string>
 
-#include "bigquery/internal/bigquerystorage_stub.h"
-#include "bigquery/internal/connection_impl.h"
-#include "bigquery/version.h"
+#include "google/cloud/bigquery/internal/bigquerystorage_stub.h"
+#include "google/cloud/bigquery/internal/connection_impl.h"
+#include "google/cloud/bigquery/version.h"
 #include "google/cloud/status_or.h"
 
 namespace bigquery {
