+++ google/cloud/bigquery/internal/connection_impl.h
@@ -17,9 +17,9 @@
 
 #include <memory>
 
-#include "bigquery/connection.h"
-#include "bigquery/internal/bigquerystorage_stub.h"
-#include "bigquery/version.h"
+#include "google/cloud/bigquery/connection.h"
+#include "google/cloud/bigquery/internal/bigquerystorage_stub.h"
+#include "google/cloud/bigquery/version.h"
 #include "google/cloud/status_or.h"
 
 namespace bigquery {
