+++ google/cloud/bigquery/internal/stream_reader.h
@@ -15,7 +15,7 @@
 #ifndef BIGQUERY_INTERNAL_STREAMING_READER_H_
 #define BIGQUERY_INTERNAL_STREAMING_READER_H_
 
-#include "bigquery/version.h"
+#include "google/cloud/bigquery/version.h"
 #include "google/cloud/optional.h"
 #include "google/cloud/status_or.h"
