+++ google/cloud/bigquery/version_info.h
